pdf-chat-system-qt/requirements.txt
# requirements.txt

PyQt6>=6.6.0
openai>=1.0.0
pymupdf>=1.23.0
python-dotenv>=1.0.0